Allow changing the display scale using an environment variable
Fix the render_view_to_framebuffer function by scaling the relative
position of subsurfaces by the buffer scale.

Hardcode the notch height of 80 pixels for OnePlus 6.

+++ src/util/render_to_texture_shader.cpp
@@ -184,7 +184,7 @@ void RenderToTextureShader::render(GLuint texture, int x, int y, size_t width, s
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
 	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 
-	glViewport(x, y, width, height);
+	glViewport(x, y, (int) width, (int) height);
 
 	glDrawElements(GL_TRIANGLES, 6, GL_UNSIGNED_INT, nullptr);
